+++ CHANGELOG.md
@@ -3,10 +3,10 @@
  * FEATURE - Dynamic selection of translator during game session
  * FEATURE - Support BingTranslate API
  * FEATURE - Support LEC Offline Power Translator 15
- * FEATURE - Enable custom implementations of translators that can be loaded dynamically
- * FEATURE - Removed support for Excite translate because it only support the 'WWW' API in Unity due to missing TLS1.2 implementation
+ * FEATURE - Enable custom implementations of translators
+ * FEATURE - Removed support for Excite translate because it only support the 'WWW' API in Unity due to missing TLS1.2 support
  * FEATURE - Updated Watson translate to v3
- * FEATURE - Support for 'romaji' as output language. Only google supports this at the  moment
+ * FEATURE - Support for 'romaji' as output language. Only google supports this at the moment
  * BUG FIX - Too many small fixes to mention
  * MISC - {GameExeName} variable can now be used in configuration of directories and files
  * MISC - Changed the way the 'Custom' endpoint works. See README for more info